Create Java 2
I misunderstood initial so I had to redo a lot. Im thinking of implementing a Java API but i need ot apply for license and i dont know how long that'll take. we also need to show what we are using it for so we cant do it for this stage. Currently using an authentication which is temporary for visualisation for the user.